requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-generativeai>=0.3.0
elevenlabs>=0.2.27
PyMuPDF>=1.23.5
ffmpeg-python>=0.2.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
moviepy>=1.0.3
aiofiles>=23.2.1
pydantic>=2.4.2
SQLAlchemy>=2.0.22